+++ src/extension.ts
@@ -2,24 +2,14 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as fs from "fs";
 import * as Path from "path";
-import { URLSearchParams } from "url";
 import * as vscode from "vscode";
+import { FileContentProvider } from "./cncView/FileContentTree";
 import { config } from "./util/config";
+import * as open from "open";
+import * as blowfish from "./util/encryption/encryption";
 
-/**
- * Get vscode config data and extension config
- * Terminals must be configured in vscode (example: "terminal.integrated.shell.linux": "/usr/bin/bash" in user settings)
- */
-// let terminal: unknown;
-// if (process.platform === "linux") {
-//     terminal = config.getVscodeParam("terminal.integrated.defaultProfile.linux");
-// } else if (process.platform === "win32") {
-//     terminal = config.getVscodeParam("terminal.integrated.defaultProfile.windows");
-// }
-
-
-const language = config.getParam("locale");
-const docuPath = config.getParam("documentation");
+let language: string;
+let docuPath: string;
 
 /** Outputchannel for the extension
  */
@@ -54,16 +44,22 @@ const nonAsciiCharacterDecorationType = vscode.window.createTextEditorDecoration
 let selectedLinesStatusBarItem: vscode.StatusBarItem;
 let currentOffsetStatusBarItem: vscode.StatusBarItem;
 
+let fileContentProvider: FileContentProvider;
+let fileContentTreeView: vscode.TreeView<vscode.TreeItem> | undefined;
+
 // package.json information
 let packageFile;
 let extensionPackage;
 
+let extContext: vscode.ExtensionContext;
 // Regular expression variables
 // Technology regex for too, feed, spindle rpm
 const regExTechnology = new RegExp("([TFS])([0-9]+)");
 // Blocknumber regex
 const regExpBlocknumbers = new RegExp(/^((\s?)((\/)|(\/[1-9]{0,2}))*?(\s*?)N[0-9]*(\s?))/);
 const regExpLabels = new RegExp(/(\s?)N[0-9]*:{1}(\s?)|\[.*\]:{1}/);
+
+//
 
 /**
  * This method is called when the extension is activated
@@ -72,16 +68,15 @@ const regExpLabels = new RegExp(/(\s?)N[0-9]*:{1}(\s?)|\[.*\]:{1}/);
  * @param {vscode.ExtensionContext} context
  */
 export function activate(context: vscode.ExtensionContext): void {
+    //save the context for dynamical command registeration
+    extContext = context;
     // Get package.json informations
     extensionPackage = Path.join(context.extensionPath, "package.json");
     packageFile = JSON.parse(fs.readFileSync(extensionPackage, "utf8"));
 
-    // enable/disable outputchannel
-    if (config.getParam("outputchannel")) {
-        outputChannel.show();
-    } else {
-        outputChannel.hide();
-    }
+    //get config params in module scope lets
+    updateConfig();
+
 
     // Output extension name and version number in console and output window ISG-CNC
     if (packageFile) {
@@ -103,6 +98,14 @@ export function activate(context: vscode.ExtensionContext): void {
             return [];
         }
     });
+
+    vscode.workspace.onDidChangeConfiguration(() => {
+        updateConfig();
+    });
+
+    //NC-file sidebar tree provider
+    const currentFile = vscode.window.activeTextEditor?.document.uri;
+    fileContentProvider = new FileContentProvider(currentFile, extContext);
 
     // commands
     context.subscriptions.push(
@@ -144,7 +147,27 @@ export function activate(context: vscode.ExtensionContext): void {
             findNonAsciiCharacters()
         )
     );
-
+    context.subscriptions.push(
+        vscode.commands.registerCommand("isg-cnc.EncryptAnyFile", () =>
+            blowfish.encryptFileFromSystem()
+        )
+    );
+    context.subscriptions.push(
+        vscode.commands.registerCommand("isg-cnc.DecryptAnyFile", () =>
+            blowfish.decryptFileFromSystem()
+        )
+    );
+    context.subscriptions.push(
+        vscode.commands.registerCommand("isg-cnc.EncryptThis", (inputUri) =>
+            blowfish.encryptThis(inputUri)
+        )
+    );
+    context.subscriptions.push(
+        vscode.commands.registerCommand("isg-cnc.DecryptThis", (inputUri) =>
+            blowfish.decryptThis(inputUri)
+        )
+    );
+    
     // add status bar items
     addSelectedLinesStatusBarItem(context);
     addCurrentOffsetStatusBarItem(context);
@@ -160,6 +183,7 @@ export function activate(context: vscode.ExtensionContext): void {
     if (activeEditor) {
         triggerUpdateDecorations();
     }
+
 
     // vscode.window.onDidChangeActiveTextEditor((editor) => {
     //     activeEditor = editor;
@@ -243,16 +267,57 @@ function addSelectedLinesStatusBarItem(context: vscode.ExtensionContext) {
     selectedLinesStatusBarItem.command = myCommandId;
     context.subscriptions.push(selectedLinesStatusBarItem);
 
-    // register some listener that make sure the status bar
-    // item always up-to-date
+    // register some listener that make sure the status bar 
+    // item and the currently opened file always up-to-date
+
     context.subscriptions.push(
-        vscode.window.onDidChangeActiveTextEditor(updateSelectedLinesStatusBarItem)
+        vscode.window.onDidChangeActiveTextEditor(activeTextEditorChanged)
     );
+    // refresh status bar and current open file once at start
+    activeTextEditorChanged();
+
     context.subscriptions.push(
         vscode.window.onDidChangeTextEditorSelection(
             updateSelectedLinesStatusBarItem
         )
     );
+}
+
+/**
+ * Updates the config parameters saved in the module-scoped lets and settings.
+ */
+function updateConfig() {
+    language = config.getParam("locale");
+    docuPath = config.getParam("documentation").split('"').join("").split('\\').join('/');
+
+
+    // enable/disable outputchannel
+    if (config.getParam("outputchannel")) {
+        outputChannel.show();
+    } else {
+        outputChannel.hide();
+    }
+}
+
+/**
+ * Will be called by onDidChangeActiveTextEditor-Listener.
+ * Updates the file-content-treeview and the selected
+ * status bar lines.
+ */
+function activeTextEditorChanged() {
+    //Tree view
+    try {
+        const currentFile = vscode.window.activeTextEditor?.document.uri;
+        fileContentProvider.updateTreeView(currentFile);
+        fileContentTreeView = vscode.window.createTreeView('cnc-show-filecontent', {
+            treeDataProvider: fileContentProvider
+        });
+    } catch (e: any) {
+        vscode.window.showErrorMessage(e);
+    }
+
+    //Status Bar
+    updateSelectedLinesStatusBarItem();
 }
 
 /**
@@ -388,7 +453,7 @@ async function goToPosition(): Promise<void> {
  *
  * @param {number} pos
  */
-function setCursorPosition(pos: number) {
+export function setCursorPosition(pos: number) {
     const { activeTextEditor } = vscode.window;
     if (activeTextEditor) {
         const { document } = activeTextEditor;
@@ -662,53 +727,21 @@ function digitCount(nr: number): number {
 }
 
 /**
- * Generate an terminal and load the html documentation in webbrowser.
- * Webbrowser location reading from extension setting browser.
- *
+ *Load the ISG-CNC Kernel html documentation in default webbrowser.
  */
 function startDocu() {
     const docuAddress = getContextbasedSite();
     outputChannel.appendLine(docuAddress);
 
-    const terminal = vscode.window.createTerminal({
-        name: "ISG-CNC",
-        hideFromUser: false
-    });
-    let args;
-    let browserPath;
-
-    if (process.platform === "linux") {
-        browserPath = `${config.getParam("browser-linux")}`;
-    } else if (process.platform === "win32") {
-        browserPath = `${config.getParam("browser-windows")}`.replace("\"", "");
-    }
-
-    if (docuAddress !== "" && docuAddress.startsWith("http")) {
-        args = docuAddress;
-    } else {
-        args = `"file://${docuAddress}"`;
-    }
-
-    // example that works:
-    // "C:\Program Files\Mozilla Firefox\firefox.exe"
-    // "file://c:/Users/Andre/Documents/%21%21%21ISG/ISG-Doku/de-DE/search.html?q=G54"
-
-    if (terminal.name !== "PowerShell") {
-        browserPath = `"${browserPath}"`;
-    } else {
-        browserPath = `& "${browserPath}"`;
-    }
-
     outputChannel.appendLine(`Path to the documentation: ${docuPath}`);
     outputChannel.appendLine(`Address to the website: ${docuAddress}`);
-    outputChannel.appendLine(`Commandpart: ${browserPath} and Argumentpart: ${args}`);
 
-    terminal.sendText(browserPath + " " + args);
+    open(docuAddress);
 }
 
 /**
  * Function to build the Address to the documentation.
- * Standard web Address is: https://www.isg-stuttgart.de/kernel-html5/
+ * Standard web Address is: https://www.isg-stuttgart.de/fileadmin/kernel/kernel-html/de-DE/index.html
  * Additional read the language from extension settings and the documentation path (local or web)
  * Returns the combined Address string.
  *
@@ -716,33 +749,37 @@ function startDocu() {
  */
 function getContextbasedSite(): string {
     let searchContext: string;
-    let docuPath: string = "";
+    let localeDocuPath: string = docuPath;
     let docuAddress: string = "";
     const { activeTextEditor } = vscode.window;
     if (activeTextEditor) {
         const { document } = activeTextEditor;
         if (document) {
-            if (docuPath !== undefined && docuPath !== "") {
-                docuPath = docuPath as string;
-                docuPath = docuPath.split('"').join("").split('\\').join('/');
-                if (!docuPath.endsWith('/')) {
-                    docuPath += "/" + `${language}/`;
+            if (localeDocuPath !== undefined && localeDocuPath !== "") {
+                if (!localeDocuPath.endsWith('/')) {
+                    localeDocuPath += "/" + `${language}/`;
                 } else {
-                    docuPath += `${language}/`;
+                    localeDocuPath += `${language}/`;
                 }
             } else {
-                docuPath = `https://www.isg-stuttgart.de/kernel-html5/${language}/`;
+                localeDocuPath = `https://www.isg-stuttgart.de/fileadmin/kernel/kernel-html/${language}/`;
             }
-            if (activeTextEditor.selection.isEmpty !== true) {
-                searchContext = activeTextEditor.document.getText(
-                    activeTextEditor.selection
-                );
-                const query = new URLSearchParams();
-                query.append("q", searchContext);
-                docuAddress = docuPath + `search.html?${query.toString()}`;
-            } else {
-                docuAddress = docuPath + "index.html";
-            }
+            //IMPORTANT: THE FOLLOWING BLOCK IS FOR OPEN DOCU WITH SELECTED SEARCHING KEYWORD
+            //THE WEBSITE IS BROKEN AT THE MOMENT SO IT WILL JUST LOAD THE GENERAL DOCU WEBSITE
+            //UNCOMMENT IF THE WEBSITE WAS FIXES AND DELETE THE ASSIGNMENT UNDER THE COMMENT:
+            //  "docuAddress = localeDocuPath + "index.html";"
+            /*  if (!activeTextEditor.selection.isEmpty) {
+                 searchContext = activeTextEditor.document.getText(
+                     activeTextEditor.selection
+                 );
+                 const query = new URLSearchParams();
+                 query.append("q", searchContext);
+                 docuAddress = localeDocuPath + `search.html?${query.toString()}`;
+             } else {
+                  docuAddress = localeDocuPath + "index.html";
+             } */
+            docuAddress = localeDocuPath + "index.html";
+
         }
     }
     return docuAddress;
@@ -993,3 +1030,4 @@ function findAllToolCalls(): any {
     };
     vscode.commands.executeCommand('workbench.action.findInFiles', params);
 }
+